# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger",
        "description": "RozetkaTmrro",
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "3adf530c9aaf28030b7b676f7330305d",
                "responses": {
                    "200": {
                        "description": "List Categories."
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "operationId": "55e6a43af9275e9cc3eb9b9012577f4a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "file"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Category."
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "delete": {
                "tags": [
                    "Category"
                ],
                "operationId": "866c61c966fc32bd9ab53cc0856dfcfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор категорії",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішне видалення категорії"
                    },
                    "404": {
                        "description": "Категорії не знайдено"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                }
            }
        },
        "/api/categories/edit/{id}": {
            "post": {
                "tags": [
                    "Category"
                ],
                "operationId": "c7dd8ffc2afd8e67c95ac314fe1d84f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор категорії",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "file"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Category."
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "login",
                "requestBody": {
                    "description": "User login data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "fe8a58801cad87e84b5a3c4cb16edce7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "lastName",
                                    "name",
                                    "phone",
                                    "image",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Category."
                    }
                }
            }
        }
    }
}